Added function comments

+++ HowitzerSimulator/src/HowitzerSimulator/ProjectileMotion.java
@@ -1,9 +1,12 @@
 package HowitzerSimulator;
-
+/**
+ * ProjectileMotion class
+ *
+ */
 public class ProjectileMotion {
     private int[] position;
-    private int velocity;
-    private int mass;
+    private double velocity;
+    private double mass;
     
     public ProjectileMotion(int[] position, int velocity, int mass) {
         this.position = position;
@@ -11,27 +14,50 @@ public class ProjectileMotion {
         this.mass = mass;
     }
     
+    /**
+     * Gets the Position of the Projectile 
+     * @return int[] position
+     */
     public int[] getPosition() {
         return position;
     }
     
+    /**
+     * Sets the Position of the Projectile
+     *      * @return int[] position 
+     */
     public void setPosition(int[] position) {
         this.position = position;
     }
     
-    public int getVelocity() {
+    /**
+     * Gets the velocity of the Projectile 
+     * @return double velocity
+     */
+    public double getVelocity() {
         return velocity;
     }
-    
-    public void setVelocity(int velocity) {
+    /**
+     * Sets the velocity of the Projectile 
+     * @param double velocity
+     */
+    public void setVelocity(double velocity) {
         this.velocity = velocity;
     }
     
-    public int getMass() {
+    /**
+     * Gets the mass of the Projectile 
+     * @return double mass
+     */
+    public double getMass() {
         return mass;
     }
     
-    public void setMass(int mass) {
+    /**
+     * Sets the mass of the Projectile 
+     * @param double mass
+     */
+    public void setMass(double mass) {
         this.mass = mass;
     }
 }
